refactor: build images from heroku/heroku:STACK-build images
Rather than try and emulate the work heroku is doing, use their upstream images as is. This simplifies the build process signficantly, and allows us to only layer the necessary on top of their images.

This was not previously possible due to missing arm64 support, which is introduced in the heroku-24 stack. Thus, this will fail for all other stacks.